Rename `DonNotUseThisSwitch` to `DeprecatedSwitch`
