+++ unime/src/lib/components/index.ts
@@ -3,6 +3,7 @@ export { default as Button } from './Button.svelte';
 export { default as Checkbox } from './Checkbox.svelte';
 export { default as CredentialList } from './credentials/CredentialList.svelte';
 export { default as Favorites } from './credentials/Favorites.svelte';
+export { default as DeprecatedSwitch } from './DeprecatedSwitch.svelte';
 export { default as History } from './history/History.svelte';
 export { default as IconMessage } from './IconMessage.svelte';
 export { default as Image } from './image/Image.svelte';
@@ -18,4 +19,3 @@ export { default as SettingsSwitch } from './SettingsSwitch.svelte';
 export { default as SettingsValueLink } from './SettingsValueLink.svelte';
 export { default as StatusIndicator } from './StatusIndicator.svelte';
 export { default as Switch } from './Switch.svelte';
-export { default as DoNotUseThisSwitch } from './DoNotUseThisSwitch.svelte';
